requirements.txt
langgraph
pandas
openpyxl
openai
langchain==0.3.25
langchain-core==0.3.65
langchain-openai==0.3.12
pydantic==2.11.1
pydantic-core==2.33.0
python-dotenv 